move branch protector code to subfolder delete branch protector documentation remove  branch protector from anghammarad enum
